Move `UserID` and `User` into module `src/user.rs`
