+++ src/models/mod.rs
@@ -3,12 +3,10 @@
 pub use category::{Category, CategoryName};
 pub use password::{PasswordHash, ValidatedPassword};
 pub use transaction::{Transaction, TransactionBuilder};
-pub use user::{User, UserID};
 
 mod category;
 mod password;
 mod transaction;
-mod user;
 
 /// Alias for the integer type used for mapping to database IDs.
 pub type DatabaseID = i64;
